+++ R/pivot_wider_pknca_results.R
@@ -15,18 +15,30 @@
 #'   of the dose, relative to the last reference dose.
 #'   5) Temporarily: CDISC denomination of PK parameters related to half-life: "LAMZNPT",
 #'   "LAMZLL", "LAMZ" Used to derive `LAMZNPT` and `LAMZMTD`.
+#' @param flag_rules A named list of flagging rules to be applied to the results. Each rule
+#' should be a list with two elements: `is.checked` (logical) indicating whether the rule
+#' should be applied, and `threshold` (numeric) specifying the threshold value for flagging.
+#' The name of each rule should correspond to a parameter in the results data.frame as a PPTESTCD
+#' (e.g., "R2ADJ", "AUCPEO", "AUCPEP", "LAMZSPN").
+#' @param extra_vars_to_keep Optional character vector of variable names to join from the
+#' concentration data to the output. Default is NULL.
 #'
 #' @returns A data frame which provides an easy overview on the results from the NCA
 #'          in each profile/subject and how it was computed lambda (half life) and the results
-#'          of the NCA parameters (cmax, AUC, AUClast)
+#'          of the NCA parameters (cmax, AUC, AUClast), including new columns `Exclude` (a
+#'          derivation from pknca$exclude), `Missing`(indicating if flag parameters are missing
+#'          from PKNCA calculation), and `flagged` (indicating if the row
+#'          is ACCEPTED, FLAGGED or MISSING based on the flagging rules).
 #'
 #' @importFrom dplyr select left_join rename mutate distinct group_by arrange ungroup
 #' @importFrom dplyr filter slice across where
 #' @importFrom tidyr pivot_wider pivot_longer
-#' @importFrom purrr pmap_chr
-#' @export
+#' @importFrom purrr pmap_chr map_lgl map2_chr keep
+#' @importFrom stringr str_detect fixed str_remove
+#' @importFrom rlang syms sym
 #'
-pivot_wider_pknca_results <- function(myres) {
+#' @export
+pivot_wider_pknca_results <- function(myres, flag_rules = NULL, extra_vars_to_keep = NULL) {
   ############################################################################################
   # Derive LAMZNPT & LAMZMTD
   # ToDo: At some point this will be integrated in PKNCA and will need to be removed//modified
@@ -153,7 +165,32 @@ pivot_wider_pknca_results <- function(myres) {
     ungroup()
 
   # Add "label" attribute to columns
-  add_label_attribute(pivoted_res, myres)
+  pivoted_res <- add_label_attribute(pivoted_res, myres)
+
+  # Add flagging columns for each rule and a general "flagged" column
+  out <- .apply_results_flags(
+    data = pivoted_res,
+    pknca_res = myres$result,
+    flag_settings = flag_rules
+  )
+
+  # If extra_vars_to_keep is provided, join these variables from the conc data
+  if (length(extra_vars_to_keep) > 0) {
+    conc_data <- myres$data$conc$data
+    # Only keep columns that exist in conc_data
+    vars_to_join <- intersect(extra_vars_to_keep, names(conc_data))
+    group_vars <- group_vars(myres$data$conc)
+    if (length(vars_to_join) > 0) {
+      out <- out %>%
+        dplyr::inner_join(
+          dplyr::select(conc_data, dplyr::any_of(c(vars_to_join, group_vars))),
+          by = intersect(names(out),  c(vars_to_join, group_vars))
+        ) %>%
+        dplyr::distinct()
+    }
+  }
+
+  out
 }
 
 #' Helper function to extract exclude values
@@ -169,4 +206,110 @@ pivot_wider_pknca_results <- function(myres) {
   unique_values <- unique(trimws(split_values))
 
   if (length(unique_values) == 0) NA_character_ else paste(unique_values, collapse = ", ")
+}
+
+#' Helper function to add "label" attribute to columns based on parameter names.
+#' @noRd
+#' @keywords internal
+add_label_attribute <- function(df, myres) {
+  mapping_vr <- myres$result %>%
+    mutate(
+      PPTESTCD_unit = case_when(
+        type_interval == "manual" ~ paste0(
+          PPTESTCD, "_", start, "-", end,
+          ifelse(PPSTRESU != "", paste0("[", PPSTRESU, "]"), "")
+        ),
+        PPSTRESU != "" ~ paste0(PPTESTCD, "[", PPSTRESU, "]"),
+        TRUE ~ PPTESTCD
+      ),
+      PPTESTCD_cdisc = translate_terms(PPTESTCD, mapping_col = "PPTESTCD", target_col = "PPTEST")
+    ) %>%
+    select(PPTESTCD_cdisc, PPTESTCD_unit) %>%
+    distinct() %>%
+    pull(PPTESTCD_cdisc, PPTESTCD_unit)
+
+  mapping_cols <- intersect(names(df), names(mapping_vr))
+  attrs <- unname(mapping_vr[mapping_cols])
+
+  df[, mapping_cols] <- as.data.frame(mapply(function(col, bw) {
+    attr(col, "label") <- bw
+    col
+  }, df[, mapping_cols], attrs, SIMPLIFY = FALSE))
+  df
+}
+
+#' Apply Flagging Logic to NCA Results
+#'
+#' @description
+#' Evaluates NCA results against defined flag settings and intervals. It checks for missing
+#' parameters that were requested, and determines the final `flagged` status
+#'  (ACCEPTED, FLAGGED, or MISSING).
+#'
+#' @param data A data frame of pivoted NCA results.
+#' with additional grouping variables merged.
+#' @param pknca_res A data frame. The results object from the
+#'  PKNCA result (e.g., `res$result`).
+#' @param flag_settings A named list of flag settings. Each element must contain
+#'   `is.checked` (logical) and `threshold` (numeric or character).
+#'
+#' @returns A data frame with updated `Exclude` and `flagged` columns.
+#' @keywords internal
+#' @noRd
+.apply_results_flags <- function(data, pknca_res, flag_settings) {
+
+  # Add flagging column in the pivoted results
+  applied_flags <- purrr::keep(flag_settings, function(x) x$is.checked)
+  flag_params <- names(applied_flags)
+
+  flag_thr <- sapply(flag_settings, FUN =  function(x) x$threshold)
+  flag_rule_msgs <- c(paste0(names(flag_settings), c(" < ", " < ", " > ", " > ", " < "), flag_thr))
+
+  flag_cols <- names(data)[formatters::var_labels(data)
+                           %in% translate_terms(flag_params, "PPTESTCD", "PPTEST")]
+
+  if (length(flag_cols) > 0) {
+
+    missing_flags <- pknca_res %>%
+      filter(PPTESTCD %in% flag_params,
+             type_interval == "main") %>%
+      mutate(is_missing = is.na(PPSTRES)) %>%
+      select(-PPSTRES, -PPSTRESU, -PPORRES, -PPORRESU, -type_interval)  %>%
+      pivot_wider(
+        names_from = PPTESTCD,
+        values_from = is_missing,
+        names_prefix = "missing_"
+      ) %>%
+      mutate(Missing = pmap_chr(across(starts_with("missing_")), .extract_missing_values)) %>%
+      select(-starts_with("missing_"), -exclude, -start_dose, -end_dose)
+
+    data <- data %>%
+      left_join(missing_flags,
+                by = intersect(names(data), names(missing_flags))) %>%
+      mutate(
+        flagged = case_when(
+          sapply(Exclude, function(x) any(str_detect(x, fixed(flag_rule_msgs)))) ~ "FLAGGED",
+          is.na(Missing) ~ "ACCEPTED",
+          TRUE ~ "MISSING"
+        )
+      ) %>%
+      ungroup()
+  }
+
+  data
+}
+
+# Helper function to extract missing values
+#' @noRd
+.extract_missing_values <- function(...) {
+  # Get the values for the current row
+  vals <- c(...)
+  # Get the names associated with the TRUE values
+  missing_names <- names(vals)[which(vals == TRUE)]
+
+  if (length(missing_names) == 0) return(NA_character_)
+
+  missing_names %>%
+    stringr::str_remove("missing_") %>%
+    paste0(" is NA") %>%
+    paste(collapse = "; ")
 }