+++ inst/shiny/modules/tab_nca.R
@@ -43,7 +43,8 @@ tab_nca_ui <- function(id) {
         nav_panel(
           "Slopes Information",
           navset_pill(
-            nav_panel("Slopes Results", reactable_ui(ns("slope_results"))),
+            nav_panel("Slopes Results", card(reactable_ui(ns("slope_results")),
+                                             class = "border-0 shadow-none")),
             nav_panel("Manual Adjustments", reactable_ui(ns("manual_slopes"))),
           )
         ),
@@ -62,7 +63,7 @@ tab_nca_ui <- function(id) {
   )
 }
 
-tab_nca_server <- function(id, pknca_data, extra_group_vars) {
+tab_nca_server <- function(id, pknca_data, extra_group_vars, settings_override) {
   moduleServer(id, function(input, output, session) {
     ns <- session$ns
 
@@ -74,19 +75,33 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
     adnca_data <- reactive(pknca_data()$conc$data)
 
     # #' NCA Setup module
-    nca_setup <- setup_server("nca_setup", adnca_data, pknca_data)
+    nca_setup <- setup_server(
+      "nca_setup",
+      adnca_data,
+      pknca_data,
+      extra_group_vars,
+      settings_override
+    )
 
     processed_pknca_data <- nca_setup$processed_pknca_data
     settings <- nca_setup$settings
 
     ratio_table <- nca_setup$ratio_table
     slope_rules <- nca_setup$slope_rules
+
     session$userData$settings <- list(
       settings =  settings,
-      slope_rules = slope_rules$manual_slopes
+      slope_rules = slope_rules
     ) # This will be saved in the results zip folder
 
-    reactable_server("manual_slopes", slope_rules$manual_slopes)
+    # This will be saved in the results zip folder
+    session$userData$settings <- settings
+    session$userData$ratio_table <- ratio_table
+    session$userData$slope_rules <- slope_rules
+
+    reactable_server("manual_slopes",
+                     reactive(slope_rules()$manual_slopes),
+                     columns = NULL)
 
     # List all irrelevant warnings to suppres in the NCA calculation
     irrelevant_regex_warnings <- c(
@@ -127,13 +142,16 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
         res <- withCallingHandlers({
           processed_pknca_data %>%
             filter_slopes(
-              slope_rules$manual_slopes(),
-              slope_rules$profiles_per_subject(),
-              slope_rules$slopes_groups(),
-              check_reasons = TRUE
+              slope_rules()$manual_slopes,
+              slope_rules()$profiles_per_subject,
+              slope_rules()$slopes_groups
             ) %>%
+            # Check if there are exclusions that contains a filled reason
+            check_valid_pknca_data() %>%
             # Perform PKNCA parameter calculations
-            PKNCA_calculate_nca() %>%
+            PKNCA_calculate_nca(
+              blq_rule = settings()$data_imputation$blq_imputation_rule
+            ) %>%
             # Add bioavailability results if requested
             add_f_to_pknca_results(settings()$bioavailability) %>%
             # Apply standard CDISC names
@@ -147,7 +165,9 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
                 purrr::map(\(x) x$threshold)
             ) %>%
             # Add parameter ratio calculations
-            calculate_table_ratios_app(ratio_table = ratio_table())
+            calculate_table_ratios(ratio_table = ratio_table()) %>%
+            # Keep only parameters requested by the user
+            remove_pp_not_requested()
         },
         warning = function(w) {
           if (!grepl(paste(irrelevant_regex_warnings, collapse = "|"),
@@ -168,33 +188,6 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
 
         log_success("NCA results calculated.")
 
-        params_requested <- c(setdiff(names(PKNCA::get.interval.cols()), c("start", "end")),
-                              settings()$bioavailability)
-        # Reshape intervals, filter
-        params_not_requested <- res$data$intervals %>%
-          # add bioavailability if requested
-          mutate(!!!rlang::set_names(TRUE, settings()$bioavailability)) %>%
-          # pivot for requested params
-          pivot_longer(
-            cols = (any_of(params_requested)),
-            names_to = "PPTESTCD",
-            values_to = "is_requested"
-          ) %>%
-          # Translate terms
-          mutate(PPTESTCD = translate_terms(PPTESTCD, "PKNCA", "PPTESTCD")) %>%
-          # Group by all identifying cols EXCEPT the impute column and the value column
-          group_by(across(c(-impute, -is_requested))) %>%
-          # If all are FALSE, any(is_requested) will be FALSE.
-          summarise(
-            is_requested = any(is_requested),
-            .groups = "drop"
-          ) %>%
-          filter(!is_requested)
-
-        # Filter for requested params based on intervals
-        res$result <- res$result %>%
-          anti_join(params_not_requested, by = intersect(names(.), names(params_not_requested)))
-
         res
       }, error = function(e) {
         log_error("Error calculating NCA results:\n{conditionMessage(e)}")
@@ -206,6 +199,11 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
       })
     }) %>%
       bindEvent(input$run_nca)
+
+    observe({
+      req(res_nca())
+      session$userData$final_units <- res_nca()$data$units
+    })
 
     #' Show slopes results
     pivoted_slopes <- reactive({
@@ -223,12 +221,10 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
     reactable_server(
       "slope_results",
       pivoted_slopes,
-      download_buttons = c("csv", "xlsx"),
-      file_name = function() paste0("NCA_Slope_Results_", Sys.Date()),
       defaultPageSize = 10,
-      showPageSizeOptions = TRUE,
-      pageSizeOptions = reactive(c(10, 50, nrow(pivoted_slopes()))),
-      style = list(fontSize = "0.75em")
+      pageSizeOptions = reactive(c(10, 25, 50, nrow(pivoted_slopes()))),
+      download_buttons = c("csv", "xlsx"),
+      file_name = function() paste0("NCA_Slope_Results_", Sys.Date())
     )
 
     #' Prepares and displays the pivoted NCA results
@@ -274,7 +270,7 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
 
   } else if (grepl("^No reason provided", msg)) {
     # Handle no reason provided erros from the calculation function.
-    msg <- paste(msg, "<br><br>Please provide the reason in Setup > Slope Selector tab.")
+    msg <- msg
 
   } else {
     # Handle unknown error