+++ inst/shiny/modules/tab_nca/setup/general_exclusions.R
@@ -67,16 +67,34 @@ general_exclusions_ui <- function(id) {
       )
     ),
     # Main concentration data table with row selection and color coding
-    reactable_ui(ns("conc_table"))
+    card(reactable_ui(ns("conc_table")), class = "border-0 shadow-none")
   )
 }
 
-general_exclusions_server <- function(id, processed_pknca_data) {
+general_exclusions_server <- function(id, processed_pknca_data, general_exclusions_override) {
   moduleServer(id, function(input, output, session) {
     ns <- session$ns
     # Store the list of manual exclusions and a counter for unique button IDs
     exclusion_list <- reactiveVal(list())
     xbtn_counter <- reactiveVal(0)
+
+    # Initalise settings override if available
+    observeEvent(general_exclusions_override(), {
+      overrides <- general_exclusions_override()
+
+      if (!is.null(overrides) && length(overrides) > 0) {
+        # Reconstruct list with new button IDs to match current session context
+        new_ids <- seq_along(overrides) + xbtn_counter()
+        rehydrated_list <- purrr::map2(overrides, new_ids, function(item, id) {
+          item$xbtn_id <- paste0("remove_exclusion_reason_", id)
+          item
+        })
+
+        # Update state
+        xbtn_counter(max(new_ids))
+        exclusion_list(rehydrated_list)
+      }
+    })
 
     # Reactive for the concentration data table rendered
     conc_data <- reactive({
@@ -91,8 +109,6 @@ general_exclusions_server <- function(id, processed_pknca_data) {
       selection = "multiple",
       onClick = "select",
       borderless = TRUE,
-      defaultPageSize = 25,
-      style = list(fontSize = "0.75em"),
       rowStyle = function(x) {
         function(index) {
           excl_indices <- unlist(lapply(exclusion_list(), function(excl) excl$rows))
@@ -186,6 +202,6 @@ general_exclusions_server <- function(id, processed_pknca_data) {
     })
 
     # Return the exclusion list as a reactive
-    list(exclusion_list = exclusion_list_for_return)
+    exclusion_list_for_return
   })
 }